requirements.txt
Django==3.1.13
djangorestframework==3.12.4
